% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Adomako_Ankomah_2017}
\alias{Adomako_Ankomah_2017}
\title{Multiplicity of infection in Mali using 24-SNP barcode data (Adomako-Ankomah
 et al., 2017)}
\format{
A dataframe with 26 columns, giving time point in the year (column
  1), multiplicity category, including number of failed assays (column 2),
  and frequencies for each of these classes for all 24 SNPs (columns 3:26).
}
\source{
\href{https://dx.plos.org/10.1371/journal.pone.0170948}{Supplementary materials}
}
\usage{
data(Adomako_Ankomah_2017)
}
\description{
Data from Adomako-Ankomah et al. (2017). Here we give a brief summary of the
data - see the original paper for full details.
\cr
\cr
A longitudinal cohort study was conducted in a seasonal and high-transmission
area of Mali, in which 500 subjects aged 1–65 years were followed for 1 year.
Blood samples were collected every 2 weeks. Multiclonality of Pf infection was
measured using a 24-SNP DNA barcoding assay at 4 time-points (two in wet season,
and two in dry season).
}
\references{
\insertRef{adomako-ankomah_high_2017}{SIMPLEGEN}
}
\keyword{datasets}
