% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bankole_2018}
\alias{Bankole_2018}
\title{24-SNP barcode data from two urban cities in Nigeria (Bankole et al., 2018)}
\format{
A dataframe with 27 columns, giving sample IDs and locations (columns
  1:2), genotyping calls at all 24 SNPs (columns 3:26), and (unknown) (column
  27). Heterozygous genotyping calls are denoted "N", and missing data is
  denoted "X".
}
\source{
\href{https://malariajournal.biomedcentral.com/articles/10.1186/s12936-018-2623-8}{Data in main paper}
  
  \href{https://tabula.technology/}{Tabula data extractor}
}
\usage{
data(Bankole_2018)
}
\description{
Data from Bankole et al. (2018). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
Samples were obtained from two major urban cities, Ibadan and Enugu in
Nigeria, both high transmission settings. DNA was extracted from dried blood
impregnated filter paper and samples were genotyped using a 24-SNP barcode.
Samples were considered polygenomic if they had more than 2 heterozygous
sites (Enugu = 5/28, Ibadan = 5/37), and monogenomic otherwise.  Data were
extracted using Tabula v1.2.1.
}
\references{
\insertRef{bankole_characterization_2018}{SIMPLEGEN}
}
\keyword{datasets}
