% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bei_2018}
\alias{Bei_2018}
\title{24-SNP barcode data at two time-points in Senegal (Bei et al., 2018)}
\format{
A list of multiple data objects:
  \itemize{
    \item \code{EIR}: estimates of the EIR at each location and year
    \item \code{barcodes}: genetic data and associated sample characteristics
    \item \code{SNP_locations}: genomic locations of SNPs (key to be used with \code{barcodes})
  }
  
  \code{EIR}: A dataframe with 3 columns, giving the time, the sampling
  location, and the estimated EIR (see original paper for details). The EIR
  in 2014 in Ndiop was recorded as "0.0" in the paper, and so has been coded
  as "<0.05" here to indicate the precision of this estimate.
  
  \code{barcodes}: A dataframe with 35 columns. Gives sample characteristics
  (columns 1:8), the estimated COI and whether this indicates a
  monogenomic/polygenomic infection (columns 9:10), the individual SNPs
  (columns 11:34) and the corresponding haplotype, if known (column 35). The
  "Location" column has been added from the original data in the paper, as
  this was indicated by colour code and so was lost in processing.
  
  \code{SNP_locations}: A dataframe that acts as a key relating the SNP codes
  present in \code{barcodes} to the corresponding genomic location.
}
\source{
\href{https://academic.oup.com/jid/article/217/4/622/4793403}{Data in main paper}
  \href{https://tabula.technology/}{Tabula data extractor}
}
\usage{
data(Bei_2018)
}
\description{
Data from Bei et al. (2018). Here we give a brief summary of the data - see
the original paper for full details.
\cr
\cr
Samples were randomly selected from longitudinal cohorts collected from
Dielmo/Ndiop, Senegal, and focus on 2 distant time-points (2001-2002, and
2014). The first time-point corresponds to a period of high transmission, and
the second a period of extremely low transmission. Samples were genotyped
using a 24-SNP barcode, and complexity of infection was estimated using the
COIL algorithm. Samples were also matched against a large database of
previously published and unpublished barcodes. For samples from the low
transmission period (2014), one of the three repeated barcode clusters (n =
6) corresponded to a parasite type (haplotype 3), observed in Thiès in both
2007 and 2010. The other 2 clusters (IP1, n = 2; IP2, n = 2) had not
previously been observed. Data were extracted using Tabula v1.2.1.
}
\references{
\insertRef{bei_dramatic_2018}{SIMPLEGEN}
}
\keyword{datasets}
