% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Briggs_2021}
\alias{Briggs_2021}
\title{Inferring local and crossborder transmission with human and mobility data, 
Kanunga District in Southwest Uganda (Briggs et al., 2019)}
\format{
A list of multiple data objects:
  \itemize{
    \item \code{final_samples_Dec}: contains allele lengths with barcodes. 
    \item \code{Kanungu_pairwise_comparison_df}: epidemiological data set with
    \code{final_samples_Dec} keys.
  }
 \code{Kanungu_pairwise_comparison_df}: Metadata for the allele sizes
 located in \code{final_samples_Dec}.
}
\source{
\href{https://malariajournal.biomedcentral.com/articles/10.1186/s12936-021-03603-7}{Supplementary file 1}
}
\usage{
data(Briggs_2021)
}
\description{
Data from Briggs et al. (2021). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
The study uses 80 randomly sampled households around a single health facility
and includes all children (6 mo - 10 years old) and at least one adult
caretaker from the households. DNA was acquired though dried blood spots, and
after two rounds of PCR amplification, the PCR products were sized using
capillary electrophoresis. Finally, allele length was calculated.
}
\references{
\insertRef{briggs_withinhousehold_2021}{SIMPLEGEN}
}
\keyword{datasets}
