% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_SIMPLEGEN_loaded}
\alias{check_SIMPLEGEN_loaded}
\title{Check that SIMPLEGEN package has loaded successfully}
\usage{
check_SIMPLEGEN_loaded()
}
\description{
Simple function to check that SIMPLEGEN package has loaded
  successfully. Prints "SIMPLEGEN loaded successfully!".
}
