% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CI_inc}
\alias{CI_inc}
\title{Get confidence intervals on incidence data}
\usage{
CI_inc(n, R, alpha = 0.05)
}
\arguments{
\item{n}{the number of events.}

\item{R}{the time at risk.}

\item{alpha}{the significance level of the confidence interval (two-tailed).}
}
\description{
Given a number of events n and a time at risk R, calculates a
  confidence interval for the incidence under a Poisson model. This is
  two-tailed with significance level alpha.
}
