% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Daniels_2013}
\alias{Daniels_2013}
\title{24-SNP barcode data showing clonal and epidemic transmission in Senegal (Daniels et al., 2013)}
\format{
A dataframe with 27 columns, giving sample characteristics (columns
  1:3), and genomic data at 24 SNPs (columns 4:27). Heterozygous genotyping
  calls are identified by "N", and missing alleles are identified by "NA".
}
\source{
\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0060780}{Supplementary materials}
}
\usage{
data(Daniels_2013)
}
\description{
Data from Daniels et al. (2013). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
Note - see \code{?Galinsky_2015} for an expanded version of this same set of
barcodes. All 529 Sample IDs in this dataset are also present in
Galinsky_2015, along with an additional 445 samples.
\cr
\cr
Samples were obtained annually from 2006–2011 from the Service de Lutte
Anti-Parasitaire (SLAP) clinic in Thies, Senegal. Samples were collected
passively; with patients over the age of 12 months admitted to this study
with self-reported acute fevers within 24 hours of visiting the clinic and no
recent anti-malarial use. Whole blood spots from 2006–2011 were preserved on
filter paper, and DNA was extracted and characterised using a 24-SNP
molecular barcode. Samples were excluded if they had missing data on more
than four SNP positions. Samples were identified as polygenomic if they had
multiallelic calls at more than one SNP position, otherwise monogenomic. It
is unclear whether the data presented is for monogenomic samples only, as one
sample (SenT081) has mixed calls at 2 sites and therefore is polygenomic, but
it seems unlikely that there was only one polygenomic sample observed in the
entire dataset.
}
\references{
\insertRef{daniels_genetic_2013}{SIMPLEGEN}
}
\keyword{datasets}
