% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Daniels_2020}
\alias{Daniels_2020}
\title{24-SNP barcode data from 11 sentinel sites in Haiti (Daniels et al., 2020)}
\format{
A dataframe with 14 columns, giving sample characteristics (columns
  1:7), barcode information (columns 8:9), whether the sample was designated
  mono/polygenomic (column 10) and information on travel history (columns
  11:14).
}
\source{
\href{https://malariajournal.biomedcentral.com/articles/10.1186/s12936-020-03439-7}{Supplementary materials}
}
\usage{
data(Daniels_2020)
}
\description{
Data from Daniels et al. (2020). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
Sample sites in Haiti included four sentinel sites in three departments
(Grand’Anse, Sud, and Nippes) from among 11 sites nationwide established for
anti-malarial molecular resistance marker surveillance. Individuals of all
ages seeking treatment at clinics located at the sentinel sites with symptoms
of malaria who also tested positive for malaria by either microscopy or rapid
diagnostic test (RDT) from March 2016 to December 2017 were considered
eligible to participate. Interviews were also conducted to collect patient
information (including age, gender, recent travel history). Genomic DNA was
isolated from dried blood spots and genotyped using a 24-SNP barcode. Samples
were designated polygenomic if multiple alleles were observed at two or more
positions. After quality filtering stages, what remained was 42 polygenomic
samples and 462 monogenomic samples.
}
\references{
\insertRef{daniels_genetic_2020-1}{SIMPLEGEN}
}
\keyword{datasets}
