% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Daniels_2020B}
\alias{Daniels_2020B}
\title{24-SNP barcode data from two trials in Zambia (Daniels et al., 2020)}
\format{
A dataframe with 34 columns, giving the sample ID, collection date,
  study type, (TODO), study batch and arm (columns 1:6), the genomic data
  over all 24 SNPs (columns 7:30), and details of missingness and designated
  mono/polyclonality (columns 31:34). Heterozygous genotyping calls are
  identified by "N", and missing alleles are identified by "X". Note, the
  original downloaded data contained 18 samples with missing genotypic data,
  which have been removed to leave the 1620 samples referred to in the main
  paper.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7416975/}{Supplementary materials}
}
\usage{
data(Daniels_2020B)
}
\description{
Data from Daniels et al. (2020). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
Rapid diagnostic test-positive samples were obtained from household-level
data collection during two community-randomized trials conducted in the same
geographical region of Southern Province of Zambia between 2012 and 2016. the
first sample set (n = 836 children younger than 6 years) was collected during
the peak malaria transmission season (April–May) in both 2012 and 2013
(baseline) as part of a community-randomized controlled trial designed to
assess the impact of three rounds of an MTAT intervention that used RDTs for
testing and artemether–lumefantrine for treatment. The second sample set (n =
784 individuals 3 months or older) was obtained from an 18-month longitudinal
cohort study, with monthly follow-up visits, conducted from December 2014 to
May 2016 (cohort). The cohort was designed to evaluate a cluster randomized
controlled trial for assessing the impact of four rounds of community-wide
MDA and household-level (focal) MDA (fMDA) with DHAp compared with that of no
mass treatment (control).
\cr
Samples were genotyped using a 24-SNP barcode. Samples were designated
polygenomic if multiple alleles were observed at two or more positions,
otherwise they were designated monogenomic. Samples with missing data at 5 or
more loci were deemed to have "failed" for the purposes of subsequent
analyses, but are included in the data anyway.
}
\references{
\insertRef{daniels_evidence_2020}{SIMPLEGEN}
}
\keyword{datasets}
