% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Daniels_2020C}
\alias{Daniels_2020C}
\title{24-SNP barcode data from Richard Toll, Senegal (Daniels et al., 2020)}
\format{
A dataframe with 30 columns, giving sample ID and year (columns 1:2),
  genomic data at 24 SNPs (columns 3:26), and details of missingness and
  designated mono/polyclonality (columns 27:30). Heterozygous genotyping
  calls are identified by "N", and missing alleles are identified by "X".
}
\source{
\href{https://malariajournal.biomedcentral.com/articles/10.1186/s12936-020-03346-x}{Supplementary materials}
}
\usage{
data(Daniels_2020C)
}
\description{
Data from Daniels et al. (2020). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
Samples were obtained from routine case investigation carried out in Richard
Toll, Senegal between September 2012 and December 2015. Rapid diagnostic
tests (RDTs) were used to diagnose malaria cases either through
facility-based passive case detection (PCD) or through reactive case
detection (RACD). A standardized questionnaire was also filled out for all
participants to collect information on basic demographic information
including travel history. RDTs were used to genotype malaria infections using
a 24-SNP barcode. Samples were designated polygenomic if multiple alleles
were observed at two or more positions, otherwise they were designated
monogenomic. Samples with missing data at 5 or more loci were deemed to have
"failed" for the purposes of subsequent analyses, but are included in the
data anyway.
}
\references{
\insertRef{daniels_genetic_2020}{SIMPLEGEN}
}
\keyword{datasets}
