% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_deme_parameters}
\alias{define_deme_parameters}
\title{Define deme parameters}
\usage{
define_deme_parameters(project, H = 1000, seed_infections = 100,
  M = 1000)
}
\arguments{
\item{project}{a SIMPLEGEN project.}

\item{H}{vector specifying human population size in each deme.}

\item{seed_infections}{vector specifying the initial number of infected
humans in each deme. Infected humans are assumed to have just been bitten,
meaning they have just entered the latent phase.}

\item{M}{vector specifying mosquito population size (strictly the number of
adult female mosquitoes) in each deme.}
}
\description{
Define the parameters of each deme in a SIMPLEGEN simulation.
  These parameters will be used when simulating epidemiolocal data.
}
