% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_demograpy}
\alias{define_demograpy}
\title{Define demographic parameters}
\usage{
define_demograpy(project, life_table = NULL)
}
\arguments{
\item{project}{a SIMPLEGEN project.}

\item{life_table}{vector specifying probability of death in each one-year age
group. Final value must be 1 to ensure a closed population. If \code{NULL}
then an age-distribution from Mali is used by default.}
}
\description{
Define demography used in SIMPLEGEN simulation. The raw input is
  a life table giving the probability of death in each one-year age group.
  This is used to derive the distribution of age of death, and the stable age
  distribution.
}
