% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_epi_parameters}
\alias{define_epi_parameters}
\title{Define epidemiological parameters}
\usage{
define_epi_parameters(project, a = 0.3, p = 0.85, mu = -log(p),
  u = 12, v = 10, g = 12, prob_infection = 0.6, prob_acute = c(1,
  0.5, 0), prob_AC = 0.2, duration_acute = dgeom(1:25, 1/5),
  duration_chronic = dgeom(1:1000, 1/200), infectivity_acute = 0.07,
  infectivity_chronic = 0.07, max_innoculations = 5)
}
\arguments{
\item{project}{a SIMPLEGEN project.}

\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on
humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. \code{mu = -log(p)} unless
otherwise specified.}

\item{u}{intrinsic incubation period. The number of days from infection to
blood-stage in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to
becoming infectious in a mosquito.}

\item{g}{lag time between human blood-stage infection and production of
gametocytes.}

\item{prob_infection}{probability a human becomes infected after being bitten
by an infected mosquito.}

\item{prob_acute}{probability an infection goes through an acute phase.}

\item{prob_AC}{probability of acute infection transitioning to chronic before
clearing, as opposed to clearing directly.}

\item{duration_acute}{vector or list specifying probability distribution of
time (in days) of acute phase of disease. If a list then the first element
specifies the distribution for the first incident of acute disease, the
second element for the second incident and so on (the final distribution is
used for all remaining incidents). If a vector then the same distribution
is used for all incidents of acute disease.}

\item{duration_chronic}{equivalent to \code{duration_acute} but for chronic
phase of disease.}

\item{infectivity_acute}{probability a mosquito becomes infected after biting
a human host in the acute phase.}

\item{infectivity_chronic}{probability a mosquito becomes infected after
biting a human host in the chronic phase.}

\item{max_innoculations}{maximum number of innoculations that an individual
can hold simultaneously.}
}
\description{
Define the epidemiological parameters of a SIMPLEGEN project.
  These parameters will be used when simulating epidemiolocal data.
}
