% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_epi_params}
\alias{define_epi_params}
\title{Define parameters of epidemiological simulation model}
\usage{
define_epi_params(
  project,
  a = 0.3,
  p = 0.85,
  mu = -log(p),
  u = 12,
  v = 10,
  g = 12,
  prob_infection = 0.6,
  prob_acute = 1,
  prob_AC = 1,
  duration_acute = dgeom(1:25, 1/5),
  duration_chronic = dgeom(1:250, 1/50),
  detectability_microscopy_acute = 1,
  detectability_microscopy_chronic = 0.1,
  detectability_PCR_acute = 1,
  detectability_PCR_chronic = 1,
  time_treatment_acute = dgeom(1:100, 1/20),
  time_treatment_chronic = dgeom(1:100, 1/20),
  treatment_seeking_mean = 0.5,
  treatment_seeking_sd = 0.1,
  duration_prophylactic = dgeom(1:25, 1/5),
  infectivity_acute = 0.07,
  infectivity_chronic = 0.07,
  max_inoculations = 5,
  H = 1000,
  seed_infections = 100,
  M = 1000,
  mig_mat = diag(1),
  life_table = life_table_Mali()
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on
humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. \code{mu = -log(p)} unless
otherwise specified.}

\item{u}{intrinsic incubation period. The number of days from infection to
blood-stage in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to
becoming infectious in a mosquito.}

\item{g}{lag time between human blood-stage infection and production of
gametocytes.}

\item{prob_infection}{probability a human becomes infected after being bitten
by an infected mosquito. If a vector then each subsequent value applies to
a subsequent infection.}

\item{prob_acute}{probability an infection passes to the acute stage, rather
than going directly to the chronic stage. If a vector then each subsequent
value applies to a subsequent infection.}

\item{prob_AC}{probability an infection in the acute stage passes to the
chronic stage prior to recovery, rather than recovering directly. If a
vector then each subsequent value applies to a subsequent infection.}

\item{duration_acute, duration_chronic}{specifies the probability distribution
of duration (in days) of acute or chronic disease. Can be a vector or a
list. If a list then the first element specifies the probability
distribution for the first incident of disease, the second element
for the second incident and so on (the final element is used for all
remaining incidents). If a vector then the same probability distribution is
used for all incidents of disease. Values are automatically
normalised to a proper probability mass distribution internally (i.e. a
distribution that sums to 1).}

\item{detectability_microscopy_acute, detectability_microscopy_chronic}{specifies the probability of parasite detection via microscopy at each day
since entering the acute or chronic stages. Can be a vector or a list. If a
list then the first element specifies the probability distribution for the
first incident of disease, the second element for the second incident and
so on (the final element is used for all remaining incidents). If a vector
then the same probability distribution is used for all incidents of
disease. Values are automatically normalised to a proper probability mass
distribution internally (i.e. a distribution that sums to 1).}

\item{detectability_PCR_acute, detectability_PCR_chronic}{equivalent to
`detectability_microscopy_acute` and `detectability_microscopy_chronic`,
but for detection via polymerase chain reaction (PCR).}

\item{time_treatment_acute, time_treatment_chronic}{the probability
distribution of time (in days since entering the stage) until considering
seeking treatment for acute or chronic infection. If the infection clears
naturally prior to this time then the host does not seek treatment,
otherwise the host seeks treatment according to their host-specific
treatment seeking parameter (see \code{treatment_seeking_alpha} and
\code{treatment_seeking_beta}). Can be a vector or a list. If a list then
the first element specifies the probability distribution for the first
incident of disease, the second element for the second incident and so on
(the final element is used for all remaining incidents). If a vector then
the same probability distribution is used for all incidents of disease.
Values are automatically normalised to a proper probability mass
distribution internally (i.e. a distribution that sums to 1).}

\item{treatment_seeking_mean, treatment_seeking_sd}{treatment seeking is
modelled in two stages; first, the time at which hosts \emph{consider}
seeking treatment is drawn from the probability distribution in
\code{time_treatment_acute} or \code{time_treatment_chronic}, and second,
whether they go ahead with treatment-seeking at this point in time depends
on their personal host-specific treatment-seeking probability. These latter
probabilities are drawn independently for each host at birth from a Beta
distribution with mean `treatment_seeking_mean` and standard deviation
`treatment_seeking_sd`. Hence, these parameters describe the average
propensity to seek treatment in the population, and the degree of
inequality in treatment seeking. Note that `treatment_seeking_sd`` must be
less than `sqrt(treatment_seeking_mean*(1 - treatment_seeking_mean))``,
otherwise the Beta distribution is undefined.}

\item{duration_prophylactic}{vector defining the probability distribution of
duration (in days) of prophylaxis following treatment.}

\item{infectivity_acute, infectivity_chronic}{the probability that a mosquito
becomes infected upon biting an acutely or chronically infective human host
at each day since entering the infective state. Can be a vector or a list.
If a list then the first element specifies the infectivity for the first
incident of acute disease, the second element for the second incident and
so on (the final element is used for all remaining incidents). If a vector
then the same infectivity is used for all episodes of acute disease.}

\item{max_inoculations}{maximum number of inoculations that an individual
can hold simultaneously.}

\item{H}{vector specifying human population size in each deme.}

\item{seed_infections}{vector specifying the initial number of infected human
hosts in each deme. Infected hosts are assumed to have just been bitten,
meaning they have just entered the latent phase.}

\item{M}{vector specifying mosquito population size (strictly the number of
adult female mosquitoes) in each deme.}

\item{mig_mat}{migration matrix specifying the daily probability of human
migration.}

\item{life_table}{vector specifying the probability of death of a human host
in each one-year age group. The final value must be 1 to ensure a closed
population. Defaults to a life table taken from Mali - see
\code{?life_table_Mali} for details.}
}
\description{
Define the parameters that will be used when simulating data
  from the inbuilt SIMPLEGEN epidemiological model. Any parameters that are
  left \code{NULL} will be set to default values.
}
