% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_epi_sampling_parameters}
\alias{define_epi_sampling_parameters}
\title{Define the outputs that are produced from the transmission model}
\usage{
define_epi_sampling_parameters(
  project,
  daily = NULL,
  sweeps = NULL,
  surveys = NULL
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{daily}{a dataframe of daily outputs.}

\item{sweeps}{a dataframe of outputs at specific time points.}

\item{surveys}{a dataframe specifying random surveys.}
}
\description{
Loads a dataframe into the SIMPLEGEN project that specifies how
  one or more samples are taken from the population. This is an important
  step in the SIMPLEGEN pipeline, as ultimately genotypes are only
  constructed for these sampled individuals. The user can specify details of
  the sampling scheme, including the type of sampling and the
  locations/timepoints. Probabilities of detection are taken into account
  when producing the final sample.
}
