% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_genetic_params}
\alias{define_genetic_params}
\title{Define genetic parameters}
\usage{
define_genetic_params(
  project,
  r = 1e-06,
  alpha = 1,
  oocyst_distribution = dpois(1:10, lambda = 2),
  hepatocyte_distribution = dpois(1:10, lambda = 5),
  contig_lengths = c(643292, 947102, 1060087, 1204112, 1343552, 1418244, 1501717,
    1419563, 1541723, 1687655, 2038337, 2271478, 2895605, 3291871)
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{r}{the rate of recombination. The expected number of base pairs in a
single recombinant block is 1/r.}

\item{alpha}{parameter dictating the skew of lineage densities. Small
values of \code{alpha} create a large skew, and hence make it likely that
an oocyst will be produced from the same parents. Large values of
\code{alpha} tend toward more even densities.}

\item{oocyst_distribution}{vector specifying the probability distribution of
each number of oocysts within the mosquito midgut.}

\item{hepatocyte_distribution}{vector specifying the probability distribution
of the number of infected hepatocytes in a human host. More broadly, this
defines the number of independent draws from the oocyst products that make
it into the host bloodstream upon a bite from an infectious mosquito.}

\item{contig_lengths}{vector of lengths (in bp) of each contig.}
}
\description{
Defines all parameters relating to the genetic model. This
  includes parameters related to recombination etc. that are used when
  simulating relatedness between lineages, as well as parameters related to
  mutation and sequencing errors etc. that are used at a later stage when
  converting relatedness into actual genotypes.
}
