% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_migration}
\alias{define_migration}
\title{Define migration parameters}
\usage{
define_migration(project, migration_matrix = matrix(1))
}
\arguments{
\item{project}{a SIMPLEGEN project.}

\item{migration_matrix}{TODO - rethink this input type to allow "trip"
migration.}
}
\description{
Define default migration parameters for SIMPLEGEN simulation.
}
