% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_params}
\alias{define_params}
\title{Define epidemiological parameters}
\usage{
define_params(project, a = 0.3, p = 0.85, mu = -log(p), u = 12,
  v = 10, g = 12, prob_infection = 0.6, prob_acute = 1,
  prob_AC = 1, duration_acute = dgeom(1:25, 1/5),
  duration_chronic = dgeom(1:25, 1/5), infectivity_acute = 0.07,
  infectivity_chronic = 0.07, max_innoculations = 5, H = 1000,
  seed_infections = 100, M = 1000, life_table = NULL)
}
\arguments{
\item{project}{a SIMPLEGEN project.}

\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on
humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. \code{mu = -log(p)} unless
otherwise specified.}

\item{u}{intrinsic incubation period. The number of days from infection to
blood-stage in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to
becoming infectious in a mosquito.}

\item{g}{lag time between human blood-stage infection and production of
gametocytes.}

\item{prob_infection}{probability a human becomes infected after being bitten
by an infected mosquito. If a vector then each value applies to each
subsequent infection.}

\item{prob_acute}{probability an infection passes to the acute stage, rather
than going directly to the chronic stage. If a vector then each value
applies to each subsequent infection.}

\item{prob_AC}{probability an infection in the acute stage passes to the
chronic stage prior to recovery, rather than recovering directly. If a
vector then each value applies to each subsequent infection.}

\item{duration_acute}{specifies the probability distribution of duration (in
days) of acute disease. Can be vector or list: if a list then the first
element specifies the distribution for the first incident of acute disease,
the second element for the second incident and so on (the final element is
used for all remaining incidents); if a vector then the same distribution
is used for all incidents of acute disease.}

\item{duration_chronic}{equivalent to \code{duration_acute} but for chronic
disease.}

\item{infectivity_acute}{specifies the probability that a mosquito becomes
infected upon biting an acutely infective human host. Values specify the
infectivity at each day since entering the infective state. Can be vector
or list: if a list then the first element specifies the infectivity for the
first incident of acute disease, the second element for the second incident
and so on (the final element is used for all remaining incidents); if a
vector then the same infectivity is used for all incidents of acute
disease.}

\item{infectivity_chronic}{equivalent to \code{infectivity_acute} but for
chronic disease.}

\item{max_innoculations}{maximum number of innoculations that an individual
can hold simultaneously.}

\item{H}{vector specifying human population size in each deme.}

\item{seed_infections}{vector specifying the initial number of infected
humans in each deme. Infected humans are assumed to have just been bitten,
meaning they have just entered the latent phase.}

\item{M}{vector specifying mosquito population size (strictly the number of
adult female mosquitoes) in each deme.}

\item{life_table}{vector specifying probability of death in each one-year age
group. Final value must be 1 to ensure a closed population. If \code{NULL}
then an age distribution from Mali is used by default.}
}
\description{
Define the epidemiological parameters of a SIMPLEGEN project.
  These parameters will be used when simulating epidemiolocal data.
}
