% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_sampling_strategy}
\alias{define_sampling_strategy}
\title{Define how samples are taken from epidemiological model}
\usage{
define_sampling_strategy(project, x)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{x}{a dataframe containing all of the following columns:
\itemize{
  \item time: the time (in days) at which samples are taken.
  \item deme: the deme from which samples are taken.
  \item case_detection: the method by which cases are identified. Either
  "active" or "passive".
  \item diagnosis: the method by which infected individuals are diagnosed.
  Either "microscopy" or "PCR".
  \item n: the number of individuals screened. Note that the actual number
  of infected individuals (and hence the number of genotypes) may be lower
  than this number.
}}
}
\description{
TODO.
}
