% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EIR_prev_hay2005}
\alias{EIR_prev_hay2005}
\title{Measures of annual EIR and prevalence within the same site}
\format{
A data frame of 130 rows and two columns. Column one, "annual_EIR" is
  the annual Entomological Inoculation Rate calculated for Plasmodium
  falciparum and column 2, "prevalence" is the parasite prevalence of
  plasmodium falciparum measured at the same site
}
\source{
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3130901/bin/NIHMS35895-supplement-S2.pdf}{Supplement S2}
}
\usage{
data(EIR_prev_hay2005)
}
\description{
Data from Hay et al. (2005) of measures of the prevalence and annual Entomological
Innoculation Rate (APfEIR) of P. falciparum in Africa between 1984 and 2004 . Data were read from Supplement S2.
APfEIR was defined as falciparum infected bites per adult per night indoors, using human biting rates that were averaged over 
one year and standardized to human bait catch equivalents on adults.
}
\examples{
data(EIR_prev_hay2005)
plot( EIR_prev_hay2005$annual_EIR, EIR_prev_hay2005$prevalence,
xlab = "Annual EIR (Pf)",
ylab = "Parasite prevalence (Pf)",
main = "Annual EIR vs Prevalence, Hay et al., 2005")


}
\references{
Hay, S., Guerra, C., Tatem, A. et al. Urbanization, malaria
  transmission and disease burden in Africa. Nat Rev Microbiol 3, 81–90
  (2005) doi:10.1038/nrmicro1069 
  (\href{https://www.ncbi.nlm.nih.gov/pubmed/15608702}{PubMed}
}
\keyword{datasets}
