% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EIRprev_beier1999}
\alias{EIRprev_beier1999}
\title{Measures of annual EIR and prevalence (Beier et al.,1999)}
\format{A data frame of x rows and two columns. Column one, "annual_EIR" is
  the annual Entomological Inoculation Rate calculated for Plasmodium
  falciparum and column 2, "prevalence" is the parasite prevalence of
  plasmodium falciparum measured at the same site}
\source{
\href{https://www.researchgate.net/profile/John_Beier/publication/12867829_Short_report_Entomologic_inoculation_rates_and_Plasmodium_falciparum_malaria_prevalence_in_Africa/links/004635142eb595e31a000000/Short-report-Entomologic-inoculation-rates-and-Plasmodium-falciparum-malaria-prevalence-in-Africa.pdf}{Research
  Gate} \href{https://github.com/ankitrohatgi/WebPlotDigitizer}{Data extractor hhg}
}
\usage{
data(EIRprev_beier1999)
}
\description{
Data from Beier et al.(1999) consisting of measures of the prevalence of Plasmodium
falciparum detected by microscopy in children under five years old and annual
Entomological Innoculation Rate (EIR) for the same year, from 31 sites in Africa. Data were read
from Figure 1 in the original article using WebPlotDigitizer [https://github.com/ankitrohatgi/WebPlotDigitizer].
}
\examples{
data(EIRprev_beier1999)
plot( EIRprev_beier1999$annual_EIR, EIRprev_beier1999$prevalence,
xlab = "Annual EIR (Pf)",
ylab = "Parasite prevalence (Pf)",
main = "Annual EIR vs Prevalence, Beier et al., 1999")

}
\references{
Beier, John C., Gerry F. Killeen, and John I. Githure.
  Entomologic inoculation rates and Plasmodium falciparum malaria prevalence
  in Africa. The American journal of tropical medicine and hygiene 61.1
  (1999): 109-113. (\href{https://www.ncbi.nlm.nih.gov/pubmed/10432066}{PubMed})
}
\keyword{datasets}
