% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Galinsky_2015}
\alias{Galinsky_2015}
\title{24-SNP barcode data showing monogenomic and polygenomic infections in Senegal
(Galinsky et al., 2015)}
\format{
A dataframe with 26 columns, giving the sample ID and collection year
  (columns 1:2), and genotype calls at all 24 SNP loci (columns 3:26).
  Heterozygous genotyping calls are identified by "N", and missing alleles
  are identified by "X".
}
\source{
\href{https://malariajournal.biomedcentral.com/articles/10.1186/1475-2875-14-4}{Supplementary materials}
}
\usage{
data(Galinsky_2015)
}
\description{
Data from Galinsky et al. (2015). Here we give a brief summary of the
data - see the original paper for full details.
\cr
\cr
Samples are an expanded set of those present in another dataset (see
\code{?Daniels_2013} for further details of sample collection). The
difference is that samples here also include those that are likely
polygenomic infections.
}
\references{
\insertRef{galinsky_coil_2015}{SIMPLEGEN}
}
\keyword{datasets}
