% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_coalescent_times}
\alias{get_coalescent_times}
\title{Get coalescent time of two lineages}
\usage{
get_coalescent_times(project, lineage_IDs, max_reps = 1000, silent = FALSE)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{lineage_IDs}{a vector of two lineage IDs with which to calculate coalescent times.}

\item{max_reps}{maximum number of steps when searching for coalescent times.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
TODO.
}
