% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_haplotype_coalescence2}
\alias{get_haplotype_coalescence2}
\title{Get genome-wide coalescence times between haplotypes}
\usage{
get_haplotype_coalescence2(
  project,
  haplo_IDs,
  sampling_time,
  generations = 3,
  silent = FALSE
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{haplo_IDs}{a vector of haplotype IDs on which to calculate coalescence
times.}

\item{sampling_time}{time at which sample taken.}

\item{generations}{number of generations back to search.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
TODO
}
