% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_haplotype_relatedness}
\alias{get_haplotype_relatedness}
\title{Get the relatedness between a series of haplotypes}
\usage{
get_haplotype_relatedness(project, haplo_IDs, generations = 3, silent = FALSE)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{haplo_IDs}{a vector of haplotype IDs on which to calculate pairwise
relatedness.}

\item{generations}{number of generations back to search.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
Calculates the relatedness up to a defined number of generations
  between all pairs of input haplotypes.
}
