% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_sample_relatedness}
\alias{get_sample_relatedness}
\title{Get the relatedness between a series of samples}
\usage{
get_sample_relatedness(project, sample_IDs, generations = 3, silent = FALSE)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{sample_IDs}{a vector of sample IDs on which to calculate pairwise
relatedness.}

\item{generations}{number of generations back to search.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
TODO
}
