% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Obaldia_2015}
\alias{Obaldia_2015}
\title{24-SNP barcode data from clonal outbreak in Panama (Obaldia et al., 2015)}
\format{
A dataframe with 30 columns, giving sample characteristics (columns
  1:5), genomic data at 24 SNPs (columns 6:29), and major group membership
  (column 30). Heterozygous genotyping calls are identified by "N", and
  missing alleles are identified by "NA".
}
\source{
\href{https://academic.oup.com/jid/article-lookup/doi/10.1093/infdis/jiu575}{Data extracted manually from paper}
}
\usage{
data(Obaldia_2015)
}
\description{
Data from Obaldia et al. (2015). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
This study focused on 37 P. falciparum isolates collected during 2003–2008
from individuals in malaria-endemic provinces in eastern Panama, along with
20 isolates collected during 2011–2012 from healthcare facilities in 3
malaria endemic sites in Colombia. Samples were sequenced using a 24-SNP
barcode. Samples were considered polyclonal if they contained multiple
alleles at two or more loci, otherwise they were considered monoclonal.
Samples were clustered into major groups using STRUCTURE software.
}
\references{
\insertRef{obaldia_clonal_2015}{SIMPLEGEN}
}
\keyword{datasets}
