% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCRMicro}
\alias{PCRMicro}
\title{Measures of prevalence by PCR and Microscopy}
\format{An object of class \code{list} of length 5.}
\source{
\href{https://github.com/cwhittaker1000/Sub_Patent_Malaria_Analysis/blob/master/Data/SI_Systematic_Review_Results_R_Import.csv}{Systematic Review Results}
}
\usage{
data(PCRMicro)
}
\description{
Data from preprint by Whittaker et al. (BioRxiv, 2020). 
Data were obtained from [https://github.com/cwhittaker1000/Sub_Patent_Malaria_Analysis/]
}
\references{
Global & Temporal Patterns of Submicroscopic Plasmodium falciparum Malaria Infection
Charles Whittaker, Hannah Slater, Teun Bousema, Chris Drakeley, Azra Ghani, Lucy Okell
bioRxiv 554311; doi: https://doi.org/10.1101/554311
https://doi.org/10.1101/554311
  (\href{https://www.ncbi.nlm.nih.gov/pubmed/15608702}{PubMed}
}
\keyword{datasets}
