% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCRmicro_age_whittaker2021}
\alias{PCRmicro_age_whittaker2021}
\title{Measures of prevalence by PCR and Microscopy, broken down by age}
\format{
A data frame of 164 rows and 21 columns. Columns are identical to the
  \code{PCRmicro_full_whittaker2021} dataset, with the addition of two
  columns giving the raw and processed age groups corresponding to each
  prevalence estimate.
}
\source{
\href{https://github.com/cwhittaker1000/submicroscopic_malaria/blob/master/Data/SI_Systematic_Review_Results_R_Import.csv}{Systematic Review Results}
}
\usage{
data(PCRmicro_age_whittaker2021)
}
\description{
Age-dissagregated data on Plasmodium falciparum prevalence by microscopy vs.
by PCR. Taken from the systematic review by Whittaker et al. (Lancet Microbe,
2021), downloaded from the associated
\href{https://github.com/cwhittaker1000/Sub_Patent_Malaria_Analysis/}{GitHub
repository}. See \code{?PCRmicro_full_whittaker2021} for more information
about this and related datasets.
}
\references{
Charles Whittaker, Hannah Slater, Rebecca Nash, Teun Bousema,
  Chris Drakeley, Azra C. Ghani, and Lucy C. Okell. Global patterns of
  submicroscopic Plasmodium falciparum malaria infection: insights from a
  systematic review and meta-analysis of population surveys. Lancet Microbe
  (2021) https://doi.org/10.1016/S2666-5247(21)00055-0
  (\href{https://www.sciencedirect.com/science/article/pii/S2666524721000550}{ScienceDirect})
}
\keyword{datasets}
