% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_age_states}
\alias{plot_age_states}
\title{Plot age distribution of states}
\usage{
plot_age_states(project, sample_time = NULL, deme = 1, state = "S")
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{sample_time}{which time-orderd sample to plot.}

\item{deme}{which deme to plot.}

\item{state}{which host state to plot on the y-axis.}
}
\description{
If project contains age distributions in the output from running
  the transmission model, then this function can be used to produce a simple
  barplot for a specified state.
}
