% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_daily_prevalence}
\alias{plot_daily_prevalence}
\title{Plot daily prevalence}
\usage{
plot_daily_prevalence(
  project,
  state = c("S", "E", "A", "C", "P"),
  diagnostic = "true",
  age_min = 0,
  age_max = 100,
  inoculations = -1
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{state, diagnostic, age_min, age_max, inoculations}{which rows of daily output to subset to.}
}
\description{
Plot the daily prevalence in a series of defined model states.
}
