% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_daily_states}
\alias{plot_daily_states}
\title{Plot daily counts of each host state}
\usage{
plot_daily_states(project, deme = 1, states = c("S", "E", "A", "C", "P"))
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{deme}{which deme to plot.}

\item{states}{which states to plot. Can be any subset of the column headings
in daily counts.}
}
\description{
TODO.
}
