% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_epi_data}
\alias{plot_epi_data}
\title{Plotting SIMPLEGEN outputs against epidemiological data}
\usage{
plot_epi_data(model_out, dataname)
}
\arguments{
\item{model_out}{list of SIMPLEGEN model outputs from running sim_epi function}

\item{dataname}{Data to compare SIMPLEGEN output to. "micro_vs_PCR" for plotting
the relationship between prevalence by microscopy and prevalence by PCR.
"prevalence_EIR" is the relationship between Annual EIR and prevalence,
"prevalence_incidence" plots the relationship between Prevalence in 2-10 year olds
 against incidence in 0-5 year olds, and the modelled relationship between these 
 from Griffin et a. 2014
 
 TODO - Need to add Griffin et al. 2014  model fit data and example list of SIMPLEGEN
 model outputs}
}
\description{
Plotting SIMPLEGEN outputs against epidemiological data
}
\examples{

plot_epi_data(model_out, "micro_vs_PCR")
plot_epi_data(model_out, "prevalence_EIR")
plot_epi_data(model_out, "prevalence_incidence")
}
