% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_epi_distribution}
\alias{plot_epi_distribution}
\title{Generic function for plotting a SIMPLEGEN flexible distribution}
\usage{
plot_epi_distribution(project, name = "duration_acute")
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{name}{the name of the distribution to be plotted, as it is found within
the \code{project$epi_parameters}}
}
\description{
Many distributions within the inbuilt SIMPLEGEN transmission
  model can be defined flexibly, as either a vector or a list of vectors.
  This function facilitates quick visualisation of these distributions.
}
