% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_treatment_seeking}
\alias{plot_treatment_seeking}
\title{Plot distribution of treatment seeking in the population}
\usage{
plot_treatment_seeking(project)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}
}
\description{
TODO.
}
