% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prev_inc_griffin2014}
\alias{prev_inc_griffin2014}
\title{Data on prevalence and incidence broken down by age, used in Griffin et al. 2014 model fit}
\format{
A dataframe of 262 rows and 15 columns. Columns are defined as follows:
  \itemize{
    \item \code{country}: country in which study was conducted.
    \item \code{reference}: the original reference for this data. All references can be found in Griffin et al. (2010) and Griffin et al. (2014).
    \item \code{site_name}: name of the study site.
    \item \code{site_index}: a unique numerical index given to each site (sites are contained within studies).
    \item \code{age0}, \code{age1}: lower and upper ends of this age group. Brackets are open on the right, for example if \code{age0 = 1} and \code{age1 = 2} then this corresponds to individuals of one year of age (rather than one and two year olds).
    \item \code{type}: whether incidence or prevalence estimate.
    \item \code{numer}, \code{denom}: numerator and denominator values. If prevalence data then denominator is total population size, if incidence data then denominator is total time at risk.
    \item \code{case_detection}: method of case detection, for example active case detection (ACD) vs. passive case detection (PCD).
    \item \code{meanEIR}, \code{sd_hi}, \code{sd_low}: estimate of the EIR in the region. Estimates are given in terms of a mean estimate and a lower and upper 95% interval. These values case be used to define a prior distribution on the EIR.
    \item \code{alpha(prop treated)}, \code{beta(prop treated)}: shape parameters of a Beta prior on treatment rates in the population. These can be used as a prior in model fitting, or the value \code{alpha / (alpha + beta)} can be used as a point estimate of the proportion of cases receiving treatment.
  }

 @keywords datasets
}
\usage{
prev_inc_griffin2014
}
\description{
Data used to fit the model described in Griffin et al. (2010) and Griffin et
al. (2014). Data consists of multiple prevalence and incidence estimates,
broken down by age and study site. Alongside observation counts (i.e.
numerator and denominator values) data contains estimated distributions of
EIR and treatment rates that can be used to define priors in model fitting.
}
\details{
@docType data
 
 @usage data(prev_inc_griffin2014)
}
\references{
Jamie T. Griffin, Deirdra Hollingsworth, Lucy C. Okell, Thomas S.
  Churcher, Michael White, Wes Hinsley, Teun Bousema, Chris J. Drakeley, Neil
  M. Ferguson, Maria-Gloria Basanes and Azra C. Ghani. Reducing Plasmodium
  falciparum Malaria Transmission in Africa: A Model-Based Evaluation of
  Intervention Strategies. PLoS Medicine (2010)
  doi:10.1371/journal.pmed.1000324
  (\href{https://pubmed.ncbi.nlm.nih.gov/20711482/}{PubMed})
  
  Jamie T. Griffin, Neil M. Ferguson and Azra C. Ghani. Estimates of the
  changing age-burden of Plasmodium falciparum malaria disease in sub-Saharan
  Africa. Nature Communications (2014) DOI: 10.1038/ncomms4136
  (\href{https://pubmed.ncbi.nlm.nih.gov/24518518/}{PubMed})
}
\keyword{datasets}
