% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prune_transmission_record}
\alias{prune_transmission_record}
\title{Prune the transmission record}
\usage{
prune_transmission_record(
  project,
  transmission_record_location = "",
  silent = FALSE
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{transmission_record_location}{the file path to a transmission record
already written to file.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
Reads in a saved transmission record from file. Combines this
  with sampling information in the project to produce a pruned version of the
  transmission record that contains only the events relevant to the final
  sample. This pruned record is saved to the project.
}
