% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{retrieve_prev}
\alias{retrieve_prev}
\title{Retrieve prevalence}
\usage{
retrieve_prev(
  data,
  case_detection,
  diagnosis,
  sampling_time,
  sampled = FALSE,
  deme = 1
)
}
\arguments{
\item{data}{Model output from SIMPLEGEN sim_epi() function epi_output$daily_values}

\item{case_detection}{Method of case detection, "Active" or "Passive"}

\item{diagnosis}{Method of diagnosis, "Microscopy" or "PCR"}

\item{sampling_time}{Sampling time, currently input should be a numeric day e.g. 100 would represent the 100th day of simulation.}

\item{sampled}{TO DO - sample from bernoulli distribution to generate simulated numbers of cases detected on each day - done in SIMPLEGEN internally already}

\item{deme}{deme of interest - currently assumes just a single deme analysed - i.e. can't pool results of multiple demes}
}
\value{
returns named vector :  c("annual_EIR", "prevalence_true", "prevalence_detected")
}
\description{
Retrieve prevalence
}
