% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Roh_2019}
\alias{Roh_2019}
\title{Microsatellite data across 26 loci, Eswatini (Roh et al., 2019)}
\format{
A data frame with 10 columns, giving sample ID and barcode
(columns 1:2), locus, locus indexes, and a combined id (columns 3:5),
peaks and fluorescence (columns 6:9), and probability of being a true allele
(column 10).
}
\source{
\href{https://academic.oup.com/jid/article/220/8/1346/5514501}{Supplementary materials}
}
\usage{
data(Roh_2019)
}
\description{
Data from Roh et al. (2019). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
The study investigates local versus imported transmissions and how that
effects the genetic diversity in Eswatini. The parameters collected in the
epidemiological model include district of residence, season, case detection,
age, sex, and patient occupation. To compare genetic diversity, an 
overlapping subset of microsatellite markers were used to compare against
other countries with lower transmission rates (it was found that Eswatini had
a greater diversity of parasites (mean $H_E$ was consistently higher)). When
comparing the local and imported cases, population level diversity did not
differentiate greatly; however, a greater portion of local cases were
monoclonal and less complex. Differences in MOI and $F_ws$ between cases were
statistically significant but had marginal effect sizes and could not be used
to discriminate cases. 
\cr
\cr
Eligible cases were identified through Eswatini's national malaria surveillance
program. Samples were genotyped using microsatellites from 26 different loci,
which showed no selection based on $F_ws$ calculations. In their study, they
excluded data points where the probability of there being an allele was less
than 95%, but these are included in the data set.
}
\references{
\insertRef{roh_high_2019}{SIMPLEGEN}
}
\keyword{datasets}
