% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_block_tree}
\alias{sim_block_tree}
\title{Simulate recombinant block tree}
\usage{
sim_block_tree(project, silent = FALSE)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
Starting with the haplotype tree stored within the project,
  simulates the relatedness of lengths of the genome from a recombination
  model.
}
