% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_haplotype_tree}
\alias{sim_haplotype_tree}
\title{Simulate haplotype tree}
\usage{
sim_haplotype_tree(project, silent = FALSE)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
Reads in the pruned transmission record from file and uses this,
  along with a specified genetic model, to simulate a tree relating
  haplotypes to one another.
}
