% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_vcf}
\alias{sim_vcf}
\title{Simulate mutations and write vcf to file}
\usage{
sim_vcf(
  project,
  output_location,
  sampling_time,
  generations = 3,
  mu = 1e-07,
  silent = FALSE
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{output_location}{where to write vcfs to file.}

\item{sampling_time}{time of sample.}

\item{generations}{how many generations to look back when computing ancestral
recombination graph.}

\item{mu}{mutation rate per base per unit time.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
TODO
}
