% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simplegen_file}
\alias{simplegen_file}
\title{Load system file}
\usage{
simplegen_file(name)
}
\arguments{
\item{name}{the name of a file within the inst/extdata folder.}
}
\description{
Load a file from within the inst/extdata folder of the SIMPLEGEN
  package. File extension must be one of .csv, .txt, or .rds.
}
