% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplegen_project.R
\name{simplegen_project}
\alias{simplegen_project}
\title{Define new SIMPLEGEN project}
\usage{
simplegen_project()
}
\description{
Define a new SIMPLEGEN project. This project will hold all
  simulation inputs and outputs for a given analysis, and is initialised with
  the default values of all parameters.
}
