% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_data_2014}
\alias{study_data_2014}
\title{Data on prevalence and incidence broken down by age, used in Griffin et al. 2014 model fit}
\format{
An object of class \code{data.frame} with 262 rows and 18 columns.
}
\usage{
study_data_2014
}
\description{
TODO - more description
}
\details{
@docType data
 
 @usage data(study_data_2014)

 @keywords datasets
}
\keyword{datasets}
