% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Taylor_2017}
\alias{Taylor_2017}
\title{93-SNP barcode data from Thai-Myanmar border (Taylor et al., 2017)}
\format{
A dataframe with 98 columns, giving sample IDs, dates and locations
  (columns 1:3), estimated clonality (columns 4:5), and genotyping calls at
  all 93 SNPs (columns 6:98). Missing genetic data is denoted "N".
}
\source{
\href{https://journals.plos.org/plosgenetics/article?id=10.1371/journal.pgen.1007065}{Supplementary materials}
}
\usage{
data(Taylor_2017)
}
\description{
Data from Taylor et al. (2017). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
Samples were obtained from Shoklo Malaria Research Unit (SMRU) clinics
spanning approximately 120km of the border between Thailand and Myanmar, and
were collected between 2001-2014. 1731 samples were genotyped using a 93-SNP
barcode. Samples with 6 or more heteroallelic genotyping outcomes were
designated multiple infections (n = 558), and all other samples were denoted
single infections (n = 1173).
}
\references{
\insertRef{taylor_quantifying_2017}{SIMPLEGEN}
}
\keyword{datasets}
