% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tessema_2019}
\alias{Tessema_2019}
\title{Microsatellite data across 26 loci, Northeast Namibia (Tessema et al., 2019)}
\format{
A data frame with 31 columns, giving case ID, Health Facility,
District, Region, Country (Columns 1:5), and microsatellite lengths (6:31).
}
\source{
\href{https://doi.org/10.7554/eLife.43510}{Supplementary file 1}
}
\usage{
data(Tessema_2019)
}
\description{
Data from Tessema et al. (2019). Here we give a brief summary of the data -
see the original paper for full details.
\cr
\cr
The study collected samples from 4643 symptomatic outpatients in northeast
Namibia (Kavango East and Zambezi) with confirmed cases via rapid tests. 3871
cases were from Kavango East (March-June 2016) from 23 different health
facilities, and 772 from Zambezi (Feb 2015 and June 2016) from six
facilities.
\cr
\cr
DNA was extracted from dried blood spots using punches and strips from rapid 
diagnostic tests. If a facility had less than 100 rapid tests, all tests were
included for genotyping. If over 100 tests, then all cases with travel history
were included for genotyping, and up to 100 cases without history were also
included. In total, 2990 samples were genotyped using microsatellite markers,
in which 2585 were used for data analysis.
}
\references{
\insertRef{tessema_using_2019}{SIMPLEGEN}
}
\keyword{datasets}
