% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_epi_sampling_parameters}
\alias{define_epi_sampling_parameters}
\title{Define the outputs that are produced from the transmission model}
\usage{
define_epi_sampling_parameters(
  project,
  daily = NULL,
  age_distributions = NULL,
  surveys = NULL
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{df_sample}{a dataframe containing all of the following columns:
\itemize{
  \item time: the time (in days) at which samples are taken.
  \item deme: the deme from which samples are taken.
  \item case_detection: the method by which cases are identified. Either
  "active" or "passive".
  \item diagnosis: the method by which infected individuals are diagnosed.
  Either "microscopy" or "PCR".
  \item n: the number of individuals screened. Note that the actual number
  of infected individuals (and hence the number of genotypes) may be lower
  than this number.
}}
}
\description{
Loads a dataframe into the SIMPLEGEN project that specifies how
  one or more samples are taken from the population. This is an important
  step in the SIMPLEGEN pipeline, as ultimately genotypes are only
  constructed for these sampled individuals. The user can specify details of
  the sampling scheme, including the type of sampling and the
  locations/timepoints. Probabilities of detection are taken into account
  when producing the final sample.
}
