% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_epi_sampling_parameters}
\alias{define_epi_sampling_parameters}
\title{Define outputs produced from the transmission model}
\usage{
define_epi_sampling_parameters(
  project,
  daily = NULL,
  sweeps = NULL,
  surveys = NULL
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{daily}{a dataframe of daily outputs.}

\item{sweeps}{a dataframe of outputs at specific time points.}

\item{surveys}{a dataframe specifying surveys to be conducted on the
population.}
}
\description{
Loads one or more dataframes into the SIMPLEGEN project that
  specify which outputs will be returned from the epidemiological model.
}
