% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_genetic_params}
\alias{define_genetic_params}
\title{Define parameters of genetic simulation model}
\usage{
define_genetic_params(project, a = 0.3)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on
humans each day.}
}
\description{
Define the parameters that will be used, alonside a pruned
  transmission record, to generate genotypes.
}
