% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{life_table_Mali}
\alias{life_table_Mali}
\title{A life table taken from Mali}
\usage{
life_table_Mali()
}
\description{
The default life table used within the inbuilt SIMPLEGEN
  transmission model. Values represent probabilities of death in one-year age
  bands.
}
