% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_treatment_seeking}
\alias{plot_treatment_seeking}
\title{Plot distribution of treatment seeking in the population}
\usage{
plot_treatment_seeking(project)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}
}
\description{
Within the SIMPLEGEN transmission model each host has its own
  treatment seeking parameter, which defines the probability of actively
  seeking treatment once disease reaches a point that this becomes necessary.
  These parameter values are drawn independently for each host from a Beta
  distribution with user-defined mean and standard deviation. This function
  plots the Beta distribution for the current project values.
}
