% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prune_transmission_record}
\alias{prune_transmission_record}
\title{Prune the transmission record}
\usage{
prune_transmission_record(project, transmission_record_location = "",
  pruned_record_location = "", overwrite_pruned_record = FALSE,
  silent = FALSE)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{transmission_record_location}{the file path to a transmission record
already written to file.}

\item{pruned_record_location}{the file path that the pruned transmission
record will be written to.}

\item{overwrite_pruned_record}{if \code{TRUE} the pruned transmission record
will overwrite any existing file by the same name. \code{FALSE} by default.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
TODO
}
