% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_epi}
\alias{sim_epi}
\title{Simulate from simple individual-based model}
\usage{
sim_epi(
  project,
  max_time = 365,
  save_transmission_record = FALSE,
  transmission_record_location = "",
  overwrite_transmission_record = FALSE,
  pb_markdown = FALSE,
  silent = FALSE
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{max_time}{run simulation for this many days.}

\item{save_transmission_record}{whether to write the transmission record to
file.}

\item{transmission_record_location}{the file path to the transmission record.}

\item{overwrite_transmission_record}{if \code{TRUE} the transmission record
will overwrite any existing file by the same name. \code{FALSE} by default.}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
Simulate from the inbuilt SIMPLEGEN transmission model. Model
  parameters are taken from the \code{epi_model_parameters} slot of the
  project, and the sampling strategy is taken from the
  \code{epi_sampling_parameters} slot. Outputs are written to the
  \code{epi_output} slot.
}
