% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_epi}
\alias{sim_epi}
\title{Simulate from simple individual-based model}
\usage{
sim_epi(project, max_time = 365, output_daily_counts = TRUE,
  output_age_distributions = TRUE, output_age_times = max_time,
  output_infection_history = FALSE, silent = FALSE)
}
\arguments{
\item{project}{a SIMPLEGEN project.}

\item{max_time}{run simulation for this many days.}

\item{output_daily_counts}{whether to output daily counts of key quantities,
such as the number of infected hosts and the EIR.}

\item{output_age_distributions}{whether to output complete age distributions}

\item{output_age_times}{a vector of times at which complete age distributions
are output.}

\item{output_infection_history}{whether to output complete infection history.}

\item{silent}{whether to write messages to console.}
}
\description{
Simulate data from a simple individual-based model of humans and
  mosquitoes. Parameters are taken from the values stored within the
  \code{sim_parameters} slot of the project, and outputs are written to the
  \code{sim_output} slot.
}
