% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_epi}
\alias{sim_epi}
\title{Simulate from simple individual-based model}
\usage{
sim_epi(
  project,
  max_time = 365,
  save_transmission_record = FALSE,
  transmission_record_location = "",
  overwrite_transmission_record = FALSE,
  output_daily_counts = TRUE,
  output_age_distributions = TRUE,
  output_age_times = max_time,
  pb_markdown = FALSE,
  silent = FALSE
)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{max_time}{run simulation for this many days.}

\item{save_transmission_record}{whether to write the transmission record to
file.}

\item{transmission_record_location}{the file path that the transmission
record will be written to.}

\item{overwrite_transmission_record}{if \code{TRUE} the transmission record
will overwrite any existing file by the same name. \code{FALSE} by default.}

\item{output_daily_counts}{whether to output daily counts of key quantities,
such as the number of infected hosts and the EIR.}

\item{output_age_distributions}{whether to output complete age distributions}

\item{output_age_times}{a vector of times at which complete age distributions
are output.}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
Simulate from the inbuilt epidemiological model. Parameters are
  taken from the \code{epi_parameters} slot of the project, and basic outputs
  are written to the \code{epi_output} slot. If a sampling strategy has been
  defined then samples will also be obtained and saved in the
  \code{sample_details} slot (see \code{?define_sampling_strategy()}).
}
