% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_relatedness}
\alias{sim_relatedness}
\title{Draw tree of genome-wide relatedness from pruned transmission record}
\usage{
sim_relatedness(project, pruned_record_location = "", silent = FALSE)
}
\arguments{
\item{project}{a SIMPLEGEN project, as produced by the
\code{simplegen_project()} function.}

\item{pruned_record_location}{the file path from which the pruned
transmission record will be read.}

\item{silent}{whether to suppress written messages to the console.}
}
\description{
Reads in the pruned transmission record from file and uses this
  information to simulate relatedness between parasite lineages from a
  genetic model. The map of relatedness output by this step can be used to
  generate genotypic information of various types.
}
\details{
Reads in the pruned transmission record and creates a new node for
  each inoculation ID. Nodes at time zero are initialised with a single
  unique lineage created de novo. In subsequent generations the children of
  any given node are known from the pruned transmission record. The lineages
  from parental nodes are sampled at random, and brought together in pairs to
  produce oocysts. The recombinant products of these oocysts are then sampled
  down to produce a new generation of lineage IDs for this node, along with
  the relative densities of each lineage.
}
