% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplegen_project.R
\name{simplegen_project}
\alias{simplegen_project}
\title{Define new SIMPLEGEN project}
\usage{
simplegen_project()
}
\description{
SIMPLEGEN works using projects, which are essentially just lists
  containing all the inputs and outputs of an analysis. Often we will pass
  the project in as both input and output to allow modification, as
  demonstrated in the examples below.
}
\examples{
# define a new project and modify this project to hold parameters of the
# epidemiological model
myproj <- simplegen_project()
}
