% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_data_2014}
\alias{study_data_2014}
\title{Details of data from Griffin et al 2014 refit}
\format{An object of class \code{data.frame} with 262 rows and 12 columns.}
\usage{
study_data_2014
}
\description{
@docType data
 
 @usage data(study_data_2014)
}
\details{
@keywords datasets
}
\keyword{datasets}
